1. ajax 통신 후, reload 제거 2. favicon.ico 추가 3. 더미데이터 지우고 전부 재실행 테스트 완료

+++ src/main/java/com/hsj/force/domain/dto/MenuUpdateDTO.java
@@ -18,10 +18,10 @@ public class MenuUpdateDTO {
     private String ingredientNo2;
     private String ingredientNo3;
     private String ingredientNo4;
-    private Integer quantity1;
-    private Integer quantity2;
-    private Integer quantity3;
-    private Integer quantity4;
+    private Double quantity1;
+    private Double quantity2;
+    private Double quantity3;
+    private Double quantity4;
     private List<MenuIngredientListDTO> ingredientQuantityList;
     private String imageOriginName;
     private String imageSaveName;
